Modify: Footer 데이터 FooterData.js 이동
