+++ src/Components/Footer/FooterData.js
@@ -1,0 +1,32 @@
+export const FOOTERLINK = [
+  {
+    id: 0,
+    key: 'notice',
+    desc: '공지사항',
+    link: '/',
+  },
+  {
+    id: 1,
+    key: 'qna',
+    desc: '1:1문의',
+    link: '/',
+  },
+  {
+    id: 2,
+    key: 'clause',
+    desc: '이용약관',
+    link: '/',
+  },
+  {
+    id: 3,
+    key: 'policy',
+    desc: '개인정보처리방침',
+    link: '/',
+  },
+  {
+    id: 4,
+    key: 'shops',
+    desc: '판매처 안내',
+    link: '/',
+  },
+]
